refactor: flatten module structure, remove threads directory
- Remove concurrency/src/threads/ directory (replaced by Backend enum)
- Move files from concurrency/src/tasks/ to concurrency/src/
- Update all imports from spawned_concurrency::tasks:: to spawned_concurrency::
- Update internal crate imports accordingly
- Update examples README to reflect current architecture

The Backend enum (Async, Blocking, Thread) now provides all the
functionality previously split between tasks and threads modules,
offering a cleaner and more unified API.
