+++ examples/bank/src/server.rs
@@ -1,11 +1,8 @@
 use std::collections::HashMap;
 
 use spawned_concurrency::{
-    messages::Unused,
-    tasks::{
-        CallResponse, GenServer, GenServerHandle,
-        InitResult::{self, Success},
-    },
+    messages::Unused, CallResponse, GenServer, GenServerHandle,
+    InitResult::{self, Success},
 };
 
 use crate::messages::{BankError, BankInMessage as InMessage, BankOutMessage as OutMessage};
