+++ examples/blocking_genserver/main.rs
@@ -2,7 +2,7 @@ use spawned_rt::tasks as rt;
 use std::time::Duration;
 use std::{process::exit, thread};
 
-use spawned_concurrency::tasks::{
+use spawned_concurrency::{
     Backend, CallResponse, CastResponse, GenServer, GenServerHandle, send_after,
 };
