+++ examples/updater/src/main.rs
@@ -9,7 +9,7 @@ mod server;
 use std::{thread, time::Duration};
 
 use server::UpdaterServer;
-use spawned_concurrency::tasks::{Backend, GenServer as _};
+use spawned_concurrency::{Backend, GenServer as _};
 use spawned_rt::tasks as rt;
 
 fn main() {
