+++ examples/updater/src/server.rs
@@ -1,11 +1,8 @@
 use std::time::Duration;
 
 use spawned_concurrency::{
-    messages::Unused,
-    tasks::{
-        send_interval, CastResponse, GenServer, GenServerHandle,
-        InitResult::{self, Success},
-    },
+    messages::Unused, send_interval, CastResponse, GenServer, GenServerHandle,
+    InitResult::{self, Success},
 };
 use spawned_rt::tasks::CancellationToken;
